requirements.txt
PyJWT==2.8.0
robotframework==7.0.1
